Release/v0.14.0

+++ src/main/java/org/alliancegenome/curation_api/services/AlleleService.java
@@ -1,5 +1,6 @@
 package org.alliancegenome.curation_api.services;
 
+import java.time.OffsetDateTime;
 import java.util.List;
 import java.util.Objects;
 import java.util.stream.Collectors;
@@ -10,7 +11,10 @@ import javax.inject.Inject;
 import javax.transaction.Transactional;
 
 import org.alliancegenome.curation_api.dao.AlleleDAO;
-import org.alliancegenome.curation_api.dao.slotAnnotations.AlleleMutationTypeSlotAnnotationDAO;
+import org.alliancegenome.curation_api.dao.slotAnnotations.alleleSlotAnnotations.AlleleFullNameSlotAnnotationDAO;
+import org.alliancegenome.curation_api.dao.slotAnnotations.alleleSlotAnnotations.AlleleMutationTypeSlotAnnotationDAO;
+import org.alliancegenome.curation_api.dao.slotAnnotations.alleleSlotAnnotations.AlleleSymbolSlotAnnotationDAO;
+import org.alliancegenome.curation_api.dao.slotAnnotations.alleleSlotAnnotations.AlleleSynonymSlotAnnotationDAO;
 import org.alliancegenome.curation_api.exceptions.ObjectUpdateException;
 import org.alliancegenome.curation_api.model.entities.Allele;
 import org.alliancegenome.curation_api.model.ingest.dto.AlleleDTO;
@@ -30,8 +34,13 @@ public class AlleleService extends BaseDTOCrudService<Allele, AlleleDTO, AlleleD
 
 	@Inject AlleleDAO alleleDAO;
 	@Inject AlleleMutationTypeSlotAnnotationDAO alleleMutationTypeDAO;
+	@Inject AlleleSymbolSlotAnnotationDAO alleleSymbolDAO;
+	@Inject AlleleFullNameSlotAnnotationDAO alleleFullNameDAO;
+	@Inject AlleleSynonymSlotAnnotationDAO alleleSynonymDAO;
 	@Inject AlleleValidator alleleValidator;
 	@Inject AlleleDTOValidator alleleDtoValidator;
+	@Inject DiseaseAnnotationService diseaseAnnotationService;
+	@Inject PersonService personService;
 
 	@Override
 	@PostConstruct
@@ -57,7 +66,8 @@ public class AlleleService extends BaseDTOCrudService<Allele, AlleleDTO, AlleleD
 		return alleleDtoValidator.validateAlleleDTO(dto);
 	}
 	
-	public void removeNonUpdatedAlleles(String taxonIds, List<String> alleleCuriesBefore, List<String> alleleCuriesAfter) {
+	@Transactional
+	public void removeOrDeprecateNonUpdatedAlleles(String taxonIds, List<String> alleleCuriesBefore, List<String> alleleCuriesAfter, String dataType) {
 		log.debug("runLoad: After: " + taxonIds + " " + alleleCuriesAfter.size());
 
 		List<String> distinctAfter = alleleCuriesAfter.stream().distinct().collect(Collectors.toList());
@@ -67,12 +77,27 @@ public class AlleleService extends BaseDTOCrudService<Allele, AlleleDTO, AlleleD
 		log.debug("runLoad: Remove: " + taxonIds + " " + curiesToRemove.size());
 
 		ProcessDisplayHelper ph = new ProcessDisplayHelper(1000);
-		ph.startProcess("Deletion of disease annotations linked to unloaded " + taxonIds + " alleles", curiesToRemove.size());
+		ph.startProcess("Deletion/deprecation of disease annotations linked to unloaded " + taxonIds + " alleles", curiesToRemove.size());
 		for (String curie : curiesToRemove) {
 			Allele allele = alleleDAO.find(curie);
 			if (allele != null) {
-				deleteAlleleSlotAnnotations(allele);
-				alleleDAO.deleteAlleleAndReferencingDiseaseAnnotations(curie);
+				List<Long> referencingDAIds = alleleDAO.findReferencingDiseaseAnnotationIds(curie);
+				Boolean anyPublicReferencingDAs = false;
+				for (Long daId : referencingDAIds) {
+					Boolean daMadePublic = diseaseAnnotationService.deprecateOrDeleteAnnotationAndNotes(daId, false, "allele");
+					if (daMadePublic)
+						anyPublicReferencingDAs = true;
+				}
+				
+				if (anyPublicReferencingDAs) {
+					allele.setUpdatedBy(personService.fetchByUniqueIdOrCreate(dataType + " allele bulk upload"));
+					allele.setDateUpdated(OffsetDateTime.now());
+					allele.setObsolete(true);
+					alleleDAO.persist(allele);
+				} else {
+					deleteAlleleSlotAnnotations(allele);
+					alleleDAO.remove(curie);
+				}
 			} else {
 				log.error("Failed getting allele: " + curie);
 			}
@@ -88,8 +113,16 @@ public class AlleleService extends BaseDTOCrudService<Allele, AlleleDTO, AlleleD
 	}
 	
 	private void deleteAlleleSlotAnnotations(Allele allele) {
-		if (CollectionUtils.isNotEmpty(allele.getAlleleMutationTypes())) {
+		if (CollectionUtils.isNotEmpty(allele.getAlleleMutationTypes()))
 			allele.getAlleleMutationTypes().forEach(amt -> {alleleMutationTypeDAO.remove(amt.getId());});
-		}
+		
+		if (allele.getAlleleSymbol() != null)
+			alleleSymbolDAO.remove(allele.getAlleleSymbol().getId());
+		
+		if (allele.getAlleleFullName() != null)
+			alleleFullNameDAO.remove(allele.getAlleleFullName().getId());
+		
+		if (CollectionUtils.isNotEmpty(allele.getAlleleSynonyms()))
+			allele.getAlleleSynonyms().forEach(as -> {alleleSynonymDAO.remove(as.getId());});
 	}
 }
